ottl

Trait PathAccessor

pub trait PathAccessor: Debug {
    // Required methods
    fn get(
        &self,
        ctx: &EvalContext,
        path: &str,
        indexes: &[IndexExpr],
    ) -> Result<Value>;
    fn set(
        &self,
        ctx: &mut EvalContext,
        path: &str,
        indexes: &[IndexExpr],
        value: &Value,
    ) -> Result<()>;
}

Trait for accessing (reading and writing) path values in the context.

The evaluator calls get and set with the path and index list; path interpretation and indexing (e.g. ["key"], [0]) are the integrator’s responsibility.

Required Methods

fn get( &self, ctx: &EvalContext, path: &str, indexes: &[IndexExpr], ) -> Result<Value>

Get the value at this path with the given indexes applied.

Path interpretation, including indexing (e.g. ["key"], [0]), is not implemented by OTTL; the integrator must implement this method. For a typical “get base value then apply indexes” implementation, use crate::helpers::apply_indexes.

fn set( &self, ctx: &mut EvalContext, path: &str, indexes: &[IndexExpr], value: &Value, ) -> Result<()>

Set the value at this path with the given indexes applied.

When indexes is empty, the implementor sets the value at the path. When indexes is non-empty (e.g. my.list[0] = x), the implementor may support updating at that index or return an error.